requirements.txt
absl-py==2.1.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
fastjsonschema==2.21.1
filelock==3.13.1
fonttools==4.55.3
fsspec==2024.2.0
grad-cam==1.5.4
grpcio==1.68.1
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupytext==1.16.6
kagglehub==0.3.6
kiwisolver==1.4.7
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mpmath==1.3.0
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.3
opencv-python==4.10.0.84
packaging==24.2
pillow==10.2.0
platformdirs==4.3.6
protobuf==5.29.2
pydeps==2.0.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
stdlib-list==0.11.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomli==2.2.1
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.1
traitlets==5.14.3
ttach==0.0.3
typing_extensions==4.9.0
urllib3==2.2.3
Werkzeug==3.1.3
